+++ src/modules/Dataloader/index.js
@@ -1,0 +1,8 @@
+/**
+ * <article 
+ *  data-loader>
+ *  <div data-loader-canvas>
+ * 
+ * </div>
+ * 
+ */